+++ testsuite/settings.py
@@ -5,6 +5,7 @@ INSTALLED_APPS = INSTALLED_APPS + (
     'hoover.contrib.twofactor',
     'django_otp',
     'django_otp.plugins.otp_totp',
+    'hoover.contrib.ratelimit',
 )
 
 MIDDLEWARE_CLASSES = MIDDLEWARE_CLASSES + (