+++ CHANGES.md
@@ -3,6 +3,7 @@
 * "External" loader that fetches documents from a URL
 * set no-cache and no-referral response headers
 * TOTP login - `hoover.contrib.twofactor`
+* rate limiting - `hoover.contrib.ratelimit`
 
 ## 0.1 (2016-04-28)
 * initial release